# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/deposits/sham-cash/qr/{currency}": {
      "get": {
        "operationId": "DepositsController_qr",
        "parameters": [
          {
            "name": "currency",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The QR image of an enabled currency"
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/deposits/sham-cash": {
      "post": {
        "operationId": "DepositsController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A UUID per attempt; a retry with the same key returns the first result",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShamCashDeposit"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created (200: a replay)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
...
    "/api/deposits/usdt": {
      "post": {
        "operationId": "DepositsController_createUsdt",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A UUID per attempt; a retry with the same key returns the first result",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsdtDeposit"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created (200: a replay)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/deposits": {
      "get": {
        "operationId": "DepositsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositPage"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/deposits/{id}": {
      "get": {
        "operationId": "DepositsController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deposit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/deposits/{id}/quote": {
      "post": {
        "operationId": "DepositsController_requote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A new 15-minute quote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/deposits/{id}/receipt": {
      "post": {
        "operationId": "DepositsController_submitReceipt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReceipt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submitted for review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/deposits/{id}/txid": {
      "post": {
        "operationId": "DepositsController_submitTxid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTxid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/deposits/{id}/cancel": {
      "post": {
        "operationId": "DepositsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
...
    "/api/admin/deposits/{id}": {
      "get": {
        "operationId": "DepositsAdminController_deposit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deposit and its review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vertex Digital API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "DepositRejectReason": {
        "type": "string",
        "enum": [
          "not_received",
          "receipt_invalid",
          "receipt_used",
          "reference_other_customer",
          "wrong_account",
          "wrong_network",
          "transfer_other_customer",
          "other"
        ]
      },
...
      "ExchangeRate": {
        "type": "string"
      },
...
      "DepositMethod": {
        "type": "string",
        "enum": [
          "sham_cash",
          "usdt_trc20",
          "usdt_bep20"
        ]
      },
...
      "Currency": {
        "type": "string",
        "enum": [
          "USD",
          "SYP"
        ]
      },
...
      "ReviewEta": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "state": {
                "type": "string",
                "enum": [
                  "open"
                ]
              },
              "minutes": {
                "type": "integer",
                "minimum": 0,
                "exclusiveMinimum": true,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "state",
              "minutes"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "state": {
                "type": "string",
                "enum": [
                  "closed"
                ]
              },
              "opensAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
              }
            },
            "required": [
              "state",
              "opensAt"
            ],
            "additionalProperties": false
          }
        ]
      },
      "CreateShamCashDeposit": {
        "type": "object",
        "properties": {
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "amountUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "currency",
          "amountUnits"
        ]
      },
      "Deposit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "method": {
            "$ref": "#/components/schemas/DepositMethod"
          },
          "status": {
            "$ref": "#/components/schemas/DepositStatus"
          },
          "referenceCode": {
            "type": "string"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "declaredAmountUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "declaredUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "quote": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DepositQuote"
              }
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "submittedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "decidedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "rateFixed": {
            "type": "boolean"
          },
          "receiptRequest": {
            "nullable": true,
            "type": "object",
            "properties": {
              "at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
              },
              "note": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "at",
              "note"
            ],
            "additionalProperties": false
          },
          "payTo": {
            "nullable": true,
            "type": "object",
            "properties": {
              "accountName": {
                "type": "string"
              },
              "accountNumber": {
                "type": "string"
              },
              "qrUrl": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "accountName",
              "accountNumber",
              "qrUrl"
            ],
            "additionalProperties": false
          },
          "eta": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewEta"
              }
            ]
          },
          "credited": {
            "nullable": true,
            "type": "object",
            "properties": {
              "usdUnits": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "receivedCurrency": {
                "$ref": "#/components/schemas/Currency"
              },
              "receivedAmountUnits": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "rate": {
                "nullable": true,
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ExchangeRate"
                  }
                ]
              }
            },
            "required": [
              "usdUnits",
              "receivedCurrency",
              "receivedAmountUnits",
              "rate"
            ],
            "additionalProperties": false
          },
          "rejection": {
            "nullable": true,
            "type": "object",
            "properties": {
              "reason": {
                "$ref": "#/components/schemas/DepositRejectReason"
              },
              "note": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "reason",
              "note"
            ],
            "additionalProperties": false
          },
          "usdt": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DepositUsdt"
              }
            ]
          }
        },
        "required": [
          "id",
          "method",
          "status",
          "referenceCode",
          "currency",
          "declaredAmountUnits",
          "declaredUsdUnits",
          "quote",
          "expiresAt",
          "createdAt",
          "submittedAt",
          "decidedAt",
          "rateFixed",
          "receiptRequest",
          "payTo",
          "eta",
          "credited",
          "rejection",
          "usdt"
        ],
        "additionalProperties": false
      },
      "DepositStatus": {
        "type": "string",
        "enum": [
          "pending",
          "submitted",
          "credited",
          "rejected",
          "expired",
          "cancelled"
        ]
      },
      "DepositQuote": {
        "type": "object",
        "properties": {
          "rateId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "rate": {
            "$ref": "#/components/schemas/ExchangeRate"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "rateId",
          "rate",
          "expiresAt"
        ],
        "additionalProperties": false
      },
      "DepositUsdt": {
        "type": "object",
        "properties": {
          "method": {
            "$ref": "#/components/schemas/UsdtMethod"
          },
          "address": {
            "type": "string"
          },
          "payAmount": {
            "type": "string"
          },
          "payAmountUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "checkStatus": {
            "$ref": "#/components/schemas/UsdtCheckStatus"
          },
          "checkError": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/UsdtCheckError"
              }
            ]
          },
          "txid": {
            "nullable": true,
            "type": "string"
          },
          "explorerUrl": {
            "nullable": true,
            "type": "string"
          },
          "confirmations": {
            "nullable": true,
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "requiredConfirmations": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991
          },
          "delayed": {
            "type": "boolean"
          },
          "receivedAmountUnits": {
            "nullable": true,
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "reviewReasons": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "amount_mismatch",
                "wrong_network",
                "sent_before_deposit"
              ]
            }
          }
        },
        "required": [
          "method",
          "address",
          "payAmount",
          "payAmountUnits",
          "checkStatus",
          "checkError",
          "txid",
          "explorerUrl",
          "confirmations",
          "requiredConfirmations",
          "delayed",
          "receivedAmountUnits",
          "reviewReasons"
        ],
        "additionalProperties": false
      },
      "UsdtMethod": {
        "type": "string",
        "enum": [
          "usdt_trc20",
          "usdt_bep20"
        ]
      },
      "UsdtCheckStatus": {
        "type": "string",
        "enum": [
          "awaiting_transfer",
          "searching",
          "confirming",
          "review",
          "done"
        ]
      },
      "UsdtCheckError": {
        "type": "string",
        "enum": [
          "not_found",
          "tx_failed",
          "not_to_store",
          "wrong_token",
          "amount_too_small",
          "txid_used"
        ]
      },
...
      "CreateUsdtDeposit": {
        "type": "object",
        "properties": {
          "method": {
            "$ref": "#/components/schemas/UsdtMethod"
          },
          "amountUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "method",
          "amountUnits"
        ]
      },
...
      "DepositPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Deposit"
            }
          },
          "nextCursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
      "SubmitReceipt": {
        "type": "object",
        "properties": {
          "rateId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        }
      },
      "SubmitTxid": {
        "type": "object",
        "properties": {
          "txid": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          }
        },
        "required": [
          "txid"
        ]
      },
...
    }
  }
}